Allow Computer to guess when Human creates code
The Computer adds a guess_code method that utilizes a mixture of algor-
ithms and strategies where the mastermind board contains four holes
and six colors.

The algorithm used here is the "Donald Knuth's Five-Guess Algorithm"
(1977) where the code breaker can solve any pattern in five moves or
fewer.

Another algorithm employed is the Swaszek strategy (1999-2000) to pick
the first element in the list as the new guess once all the elements in
the secret code set have been removed.